Cleanup of datasource binding
