+++ CSHTML5.Telerik.Charts/CSHTML5.Extensions.TelerikUI.Controls.Chart/Telerik.Windows.Controls.Chart/Telerik.Windows.Controls.ChartView/Series/ChartSeries.cs
@@ -2,13 +2,9 @@
 //-------------- USINGS ---------------//
 //-------------------------------------//
 using Telerik.Charting;
-using Telerik.Windows.Controls.Primitives;
-using System.Windows.Controls;
 using System.Windows;
-using System;
 using System.Windows.Markup;
 using System.Collections;
-using System.Collections.ObjectModel;
 using System.Collections.Specialized;
 //-------------------------------------//
 //-------------------------------------//
@@ -85,8 +81,17 @@ namespace Telerik.Windows.Controls.ChartView
             }
             set
             {
+                // remove event listeneres on the old datasource
+                var oldDataSource = _dataSource;
+                INotifyCollectionChanged notifyingDataSource = oldDataSource as INotifyCollectionChanged;
+                if (notifyingDataSource != null)
+                {
+                    notifyingDataSource.CollectionChanged -= new NotifyCollectionChangedEventHandler(NotifyingDataSource_CollectionChanged);
+                }
+
+                // set to the new datasource and notify change
                 _dataSource = value;
-                INotifyCollectionChanged notifyingDataSource = _dataSource as INotifyCollectionChanged;
+                notifyingDataSource = _dataSource as INotifyCollectionChanged;
                 if (notifyingDataSource != null)
                 {
                     notifyingDataSource.CollectionChanged += new NotifyCollectionChangedEventHandler(NotifyingDataSource_CollectionChanged);
@@ -111,7 +116,6 @@ namespace Telerik.Windows.Controls.ChartView
         {
             ParentChart.Refresh();
         }
-
         //-------------------------------------//
         //-------------------------------------//
         //-------------------------------------//
